[feat] 플레이어 레벨업 추가

+++ Assets/Scripts/Monster/MonsterData.cs
@@ -8,7 +8,7 @@ public class MonsterInfo
     public int MonsterID;
     public string Name;
     public int Level;
-    public float Exp;
+    public int Exp;
     public bool AtkStance;
     public float HP;
     public float Damage;
